+++ system/src/Grav/Framework/File/Formatter/MarkdownFormatter.php
@@ -33,6 +33,8 @@ class MarkdownFormatter implements FormatterInterface
      */
     public function getFileExtension()
     {
+        user_error(__CLASS__ . '::' . __FUNCTION__ . '() is deprecated since Grav 1.5, use getDefaultFileExtension() method instead', E_USER_DEPRECATED);
+
         return $this->getDefaultFileExtension();
     }