+++ system/src/Grav/Framework/File/Formatter/YamlFormatter.php
@@ -34,6 +34,8 @@ class YamlFormatter implements FormatterInterface
      */
     public function getFileExtension()
     {
+        user_error(__CLASS__ . '::' . __FUNCTION__ . '() is deprecated since Grav 1.5, use getDefaultFileExtension() method instead', E_USER_DEPRECATED);
+
         return $this->getDefaultFileExtension();
     }
 
@@ -58,13 +60,13 @@ class YamlFormatter implements FormatterInterface
     /**
      * {@inheritdoc}
      */
-    public function encode($data)
+    public function encode($data, $inline = null, $indent = null)
     {
         try {
             return (string) YamlParser::dump(
                 $data,
-                $this->config['inline'],
-                $this->config['indent'],
+                $inline ? (int) $inline : $this->config['inline'],
+                $indent ? (int) $indent : $this->config['indent'],
                 YamlParser::DUMP_EXCEPTION_ON_INVALID_TYPE
             );
         } catch (DumpException $e) {